+++ include/mp++/detail/type_traits.hpp
@@ -153,42 +153,58 @@ using is_ncrvr = conjunction<std::is_rvalue_reference<T>, negation<std::is_const
 // we will augment them with non-standard types defined
 // on some compilers.
 template <typename T>
-struct is_integral
-    : std::integral_constant<
-          bool, disjunction<std::is_integral<T>
+using is_integral = disjunction<std::is_integral<T>
 #if defined(MPPP_HAVE_GCC_INT128)
-                            ,
-                            // NOTE: for many of these type traits, the result must hold regardless
-                            // of the cv qualifications of T. Hence, we remove them.
-                            // http://eel.is/c++draft/meta.unary.cat
-                            std::is_same<remove_cv_t<T>, __int128_t>, std::is_same<remove_cv_t<T>, __uint128_t>
+                                ,
+                                // NOTE: for many of these type traits, the result must hold regardless
+                                // of the cv qualifications of T. Hence, we remove them.
+                                // http://eel.is/c++draft/meta.unary.cat
+                                std::is_same<remove_cv_t<T>, __int128_t>, std::is_same<remove_cv_t<T>, __uint128_t>
 #endif
-                            >::value> {
-};
+                                >;
 
 template <typename T>
-struct is_signed : std::integral_constant<bool, disjunction<std::is_signed<T>
+using is_signed = disjunction<std::is_signed<T>
 #if defined(MPPP_HAVE_GCC_INT128)
-                                                            ,
-                                                            std::is_same<remove_cv_t<T>, __int128_t>
+                              ,
+                              std::is_same<remove_cv_t<T>, __int128_t>
 #endif
-                                                            >::value> {
-};
+                              >;
 
 template <typename T>
-struct is_unsigned : std::integral_constant<bool, disjunction<std::is_unsigned<T>
+using is_unsigned = disjunction<std::is_unsigned<T>
 #if defined(MPPP_HAVE_GCC_INT128)
-                                                              ,
-                                                              std::is_same<remove_cv_t<T>, __uint128_t>
+                                ,
+                                std::is_same<remove_cv_t<T>, __uint128_t>
 #endif
-                                                              >::value> {
-};
+                                >;
 
 // make_unsigned machinery,
 template <typename T, typename = void>
 struct make_unsigned_impl {
     using type = typename std::make_unsigned<T>::type;
 };
+
+// NOTE: before GCC 4.9.1 the specialisation of std::make_unsigned for wchar_t
+// is broken. See the bugreport here:
+// https://gcc.gnu.org/bugzilla/show_bug.cgi?format=multiple&id=60326
+//
+// It seems bits are also missing for char16_t and char32_t, but apparently
+// not in relation to make_unsigned (only make_signed, which we don't use
+// so far). This workaround is lifted directly from the commit cited
+// in the bugreport.
+#if __GNUC__ == 4 && (__GNUC_MINOR__ < 9 || (__GNUC_MINOR__ == 9 && __GNUC_PATCHLEVEL__ < 1))
+
+#if defined(_GLIBCXX_USE_WCHAR_T) && !defined(__WCHAR_UNSIGNED__)
+
+template <>
+struct make_unsigned_impl<wchar_t, void> {
+    using type = typename make_unsigned_impl<__WCHAR_TYPE__>::type;
+};
+
+#endif
+
+#endif
 
 #if defined(MPPP_HAVE_GCC_INT128)