+++ test/rational_basic.cpp
@@ -45,7 +45,7 @@ using sizes = std::tuple<std::integral_constant<std::size_t, 1>, std::integral_c
                          std::integral_constant<std::size_t, 10>>;
 
 using int_types = std::tuple<char, signed char, unsigned char, short, unsigned short, int, unsigned, long,
-                             unsigned long, long long, unsigned long long
+                             unsigned long, long long, unsigned long long, wchar_t
 #if defined(MPPP_HAVE_GCC_INT128)
                              ,
                              __int128_t, __uint128_t
@@ -110,7 +110,6 @@ struct int_ctor_tester {
         REQUIRE((std::is_constructible<rational, bool &&>::value));
         REQUIRE((lex_cast(rational{false}) == "0"));
         REQUIRE((lex_cast(rational{true}) == "1"));
-        REQUIRE((!std::is_constructible<rational, wchar_t>::value));
         REQUIRE((!std::is_constructible<rational, no_const>::value));
         std::cout << "n static limbs: " << S::value << ", size: " << sizeof(rational) << '\n';
         // Testing for the ctor from int_t.
@@ -127,7 +126,6 @@ struct int_ctor_tester {
         REQUIRE((std::is_constructible<rational, integer, integer>::value));
         REQUIRE((std::is_constructible<rational, integer, int>::value));
         REQUIRE((std::is_constructible<rational, short, integer>::value));
-        REQUIRE((!std::is_constructible<rational, short, wchar_t>::value));
         REQUIRE((!std::is_constructible<rational, std::string, integer>::value));
         REQUIRE((!std::is_constructible<rational, float, integer>::value));
         REQUIRE((!std::is_constructible<rational, integer, float>::value));
@@ -487,7 +485,6 @@ struct copy_move_tester {
     {
         using rational = rational<S::value>;
         using integer = typename rational::int_t;
-        REQUIRE((!std::is_assignable<rational &, const wchar_t &>::value));
         REQUIRE((!std::is_assignable<rational &, const std::vector<int> &>::value));
         REQUIRE((!std::is_assignable<const rational &, int>::value));
         rational q;
@@ -1075,7 +1072,6 @@ struct int_convert_tester {
         REQUIRE(roundtrip_conversion<rational>(true));
         REQUIRE(roundtrip_conversion<rational>(false));
         // Extra.
-        REQUIRE((!is_convertible<rational, wchar_t>::value));
         REQUIRE((!is_convertible<rational, no_conv>::value));
         // Conversion to int_t.
         integer rop;